Moving SafeNativeMethods to Interop.Windows.Kernel32 namespace, migrating usages
